Enum auteur_controlling::controller::Command

pub enum Command {
    CreateSource {
        id: String,
        uri: String,
    },
    CreateDestination {
        id: String,
        family: DestinationFamily,
    },
    CreateMixer {
        id: String,
        config: Option<HashMap<String, Value>>,
    },
    Connect {
        link_id: String,
        src_id: String,
        sink_id: String,
        config: Option<HashMap<String, Value>>,
    },
    Start {
        id: String,
        cue_time: Option<DateTime<Utc>>,
        end_time: Option<DateTime<Utc>>,
    },
    Reschedule {
        id: String,
        cue_time: Option<DateTime<Utc>>,
        end_time: Option<DateTime<Utc>>,
    },
    Remove {
        id: String,
    },
    Disconnect {
        link_id: String,
    },
    GetInfo {
        id: Option<String>,
    },
    AddControlPoint {
        controllee_id: String,
        property: String,
        control_point: ControlPoint,
    },
    RemoveControlPoint {
        id: String,
        controllee_id: String,
        property: String,
    },
}

Command variants

Variants

CreateSource

Create a source

No slot properties are available.

Fields of CreateSource

id: String

Unique identifier of the source

uri: String

URI to play back

CreateDestination

Create a destination

No slot properties are available.

Fields of CreateDestination

id: String

Unique identifier of the destination

family: DestinationFamily

Type of the destination

CreateMixer

Create a mixer

Available slot properties are all those exposed by audiomixer and compositor pads, accessible through a audio:: or video:: prefix respectively, eg video::width.

Fields of CreateMixer

id: String

Unique identifier of the mixer

config: Option<HashMap<String, Value>>

Initial configuration of the mixer

Available settings:

width, i32, 1 -> 2147483647, default 1920, controllable

height, i32, 1 -> 2147483647, default 1920, controllable

sample-rate, i32, 1 -> 2147483647, default 48000

fallback-image, String, default “”

fallback-timeout (ms), i32, 1 -> 2147483647, default 500, controllable

Connect

Connect a producer with a consumer

Fields of Connect

link_id: String

Unique identifier of the created connection

src_id: String

Identifier of an existing producer

sink_id: String

Identifier of an existing consumer

config: Option<HashMap<String, Value>>

Initial configuration of the consumer slot

Check out the documentation for the node creation function for more information on the available consumer slot config keys.

Start

Schedule any node for starting, possibly immediately

Fields of Start

id: String

Unique identifier of the node

cue_time: Option<DateTime<Utc>>

When to start the node. Immediate if None

end_time: Option<DateTime<Utc>>

When to stop the node. Never if None

Reschedule

Reschedule any node

Fields of Reschedule

id: String

Identifier of an existing node

cue_time: Option<DateTime<Utc>>

When to start the node, the current time is left unchanged if None

end_time: Option<DateTime<Utc>>

When to stop the node, the current time is left unchanged if None

Remove

Remove a node

Fields of Remove

id: String

Identifier of an existing node

Disconnect

Remove a connection between two nodes

Fields of Disconnect

link_id: String

Identifier of an existing connection

GetInfo

Retrieve the info of one or all nodes

Fields of GetInfo

id: Option<String>

The id of an existing node, or None, in which case the info of all nodes in the system will be gathered

AddControlPoint

Control a property on a node or slot

Fields of AddControlPoint

controllee_id: String

Identifier of an existing node or slot

property: String

Name of the controlled property

control_point: ControlPoint

The control point that should be added

RemoveControlPoint

Remove a previously-created control point

Fields of RemoveControlPoint

id: String

Unique identifier of the control point

controllee_id: String

Identifier of an existing node or slot

property: String

Name of the controlled property