Enum auteur_controlling::controller::GraphCommand

pub enum GraphCommand {
    CreateSource {
        id: String,
        uri: String,
    },
    CreateDestination {
        id: String,
        family: DestinationFamily,
    },
    CreateMixer {
        id: String,
        config: MixerConfig,
    },
    Connect {
        link_id: String,
        src_id: String,
        sink_id: String,
    },
    Start {
        id: String,
        cue_time: Option<DateTime<Utc>>,
        end_time: Option<DateTime<Utc>>,
    },
    Reschedule {
        id: String,
        cue_time: Option<DateTime<Utc>>,
        end_time: Option<DateTime<Utc>>,
    },
    Remove {
        id: String,
    },
    Disconnect {
        link_id: String,
    },
    GetInfo {
        id: Option<String>,
    },
}

Generic commands for creating and removing nodes, managing connections and rescheduling

Variants

CreateSource

Create a source

Fields of CreateSource

id: String

Unique identifier of the source

uri: String

URI to play back

CreateDestination

Create a destination

Fields of CreateDestination

id: String

Unique identifier of the destination

family: DestinationFamily

Type of the destination

CreateMixer

Create a mixer

Fields of CreateMixer

id: String

Unique identifier of the mixer

config: MixerConfig

Initial configuration of the mixer

Connect

Connect a producer with a consumer

Fields of Connect

link_id: String

Unique identifier of the created connection

src_id: String

Identifier of an existing producer

sink_id: String

Identifier of an existing consumer

Start

Schedule any node for starting, possibly immediately

Fields of Start

id: String

Unique identifier of the node

cue_time: Option<DateTime<Utc>>

When to start the node. Immediate if None

end_time: Option<DateTime<Utc>>

When to stop the node. Never if None

Reschedule

Reschedule any node

Fields of Reschedule

id: String

Identifier of an existing node

cue_time: Option<DateTime<Utc>>

When to start the node, the current time is left unchanged if None

end_time: Option<DateTime<Utc>>

When to stop the node, the current time is left unchanged if None

Remove

Remove a node

Fields of Remove

id: String

Identifier of an existing node

Disconnect

Remove a connection between two nodes

Fields of Disconnect

link_id: String

Identifier of an existing connection

GetInfo

Retrieve the info of one or all nodes

Fields of GetInfo

id: Option<String>

The id of an existing node, or None, in which case the info of all nodes in the system will be gathered