Module auteur_controlling::controller

Definition of the service protocol

WARNING: unstable

Structs

ControllerMessage
Messages sent from the controller to the server.

DestinationInfo
Destination-specific information

Info
A map of node-specific information in reply to a GetInfo command

MixerConfig
Configuration of a mixer’s output stream

MixerInfo
Mixer-specific information

MixerSlotInfo
Mixer-slot-specific information

NodeCommand
Node-specific commands

ServerMessage
Messages sent from the the server to the controller.

SourceInfo
Source-specific information

Enums

Command
Command variants

CommandResult
Messages sent from the the server to the controller.

DestinationCommand
Commands to execute on a destination

DestinationFamily
The available types of destinations

GraphCommand
Generic commands for creating and removing nodes, managing connections and rescheduling

MixerCommand
Commands to execute on a mixer

NodeCommands
Node-specific command variants

NodeInfo
Info variants

SourceCommand
Commands to execute on a source

State
The state of a node