Module auteur_controlling::controller

Definition of the service protocol

WARNING: unstable

Structs

ControlPoint
A property control point

ControllerMessage
Messages sent from the controller to the server.

DestinationInfo
Destination-specific information

Info
A map of node-specific information in reply to a GetInfo command

MixerInfo
Mixer-specific information

MixerSlotInfo
Mixer-slot-specific information

ServerMessage
Messages sent from the the server to the controller.

SourceInfo
Source-specific information

Enums

Command
Command variants

CommandResult
Messages sent from the the server to the controller.

ControlMode
Defines how a property should be controlled

DestinationFamily
The available types of destinations

NodeInfo
Info variants

State
The state of a node